# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ludus UX API
  description: |
    REST API for the Ludus UX web application — a Next.js frontend for managing
    [Ludus](https://ludus.cloud) cyber range infrastructure, including VM ranges,
    content sources (sync / force re-install), templates, user management,
    VNC/SPICE console access, GOAD (Game of Active Directory) lab orchestration,
    and LudusHound BloodHound→Ludus deploys.

    ## Authentication

    All protected endpoints require a valid session cookie set by `POST /api/auth/login`.
    The cookie is AES-256-GCM encrypted and HTTP-only. Sessions expire after 24 hours.

    **Admin impersonation headers** (admin sessions only):
    - `X-Impersonate-As` — target username
    - `X-Impersonate-Apikey` — target user's Ludus API key
    - `X-Ludus-Admin: true` — route to admin API port (8081)
    - `X-Ludus-User` — override Ludus user context
  # Keep in sync with package.json / APP_VERSION (src/lib/changelog.ts).
  version: 1.4.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Ludus UX application server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ludus_session
      description: AES-256-GCM encrypted HTTP-only session cookie set by the login endpoint.
...
paths:
...
  /api/templates/add:
    post:
      tags: [Templates]
      summary: Install templates on the Ludus server
      description: |
        Fetches template files from GitLab, copies them to the Ludus server via
        host SSH, and registers them with `ludus templates add`. Uses host SSH
        credentials from the application settings.
      operationId: addTemplates
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                templates:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        description: Packer `vm_name` registered with Ludus. The on-disk directory is the last segment of `path`.
                      path:
                        type: string
                        description: Relative path in the repository (e.g. `templates/debian13`). The last segment is the on-disk directory.
                      apiBase:
                        type: string
                        format: uri
                        description: GitLab repository API base URL.
                      ref:
                        type: string
                        description: Git ref (branch/tag/SHA).
                    required: [name, path, apiBase, ref]
              required: [templates]
      responses:
        '200':
          description: Installation results per template.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        success:
                          type: boolean
                        message:
                          type: string
                      required: [name, success, message]

  # --------------------------------------------------------------------------
  # Sources (Ludus ≥2.2.0)
  # --------------------------------------------------------------------------
...